Rename Project Name.
